requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
pydantic
pydantic-settings
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
python-multipart